Upgrade everything
* Upgrade all packages, and fix all TS issues.
* Switch from `next-images` to `next-optimized-images` because the
  former has a bug with SVGs
* Use (s)css-modules and remove `@zeit/next-sass`.
* Add a custom 404 page, which can be statically optimized
