+++ package.json
@@ -17,7 +17,7 @@
     "build": "yarn lint && rm -f public/themes/*.min.css && next build",
     "build-docs": "bash scripts/update-docs.sh",
     "dev": "next",
-    "lint": "eslint --cache 'src/**/*.{ts,tsx,js}' --max-warnings 0",
+    "lint": "tsc && eslint --cache 'src/**/*.{ts,tsx,js}' --max-warnings 0",
     "start": "next start"
   },
   "repository": {
@@ -28,32 +28,33 @@
     "url": "https://github.com/elastic/next-eui-starter/issues"
   },
   "dependencies": {
-    "@elastic/datemath": "^5.0.2",
-    "@elastic/eui": "18.2.1",
-    "@next/bundle-analyzer": "^9.3.1",
+    "@elastic/datemath": "^5.0.3",
+    "@elastic/eui": "24.0.0",
+    "@next/bundle-analyzer": "^9.4.2",
     "glob": "^7.1.6",
-    "moment": "^2.24.0",
-    "next": "^9.3.2",
-    "next-images": "^1.3.1",
-    "node-sass": "^4.13.1",
+    "imagemin-mozjpeg": "^8.0.0",
+    "imagemin-optipng": "^7.1.0",
+    "moment": "^2.26.0",
+    "next": "^9.4.2",
+    "next-optimized-images": "^2.6.0",
+    "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1"
   },
   "devDependencies": {
-    "@types/node": "^13.9.3",
-    "@typescript-eslint/eslint-plugin": "^2.25.0",
-    "@typescript-eslint/parser": "^2.25.0",
-    "@zeit/next-sass": "^1.0.1",
-    "copy-webpack-plugin": "^5.1.1",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.10.1",
+    "@types/node": "^14.0.4",
+    "@typescript-eslint/eslint-plugin": "^2.34.0",
+    "@typescript-eslint/parser": "^2.34.0",
+    "copy-webpack-plugin": "^6.0.1",
+    "eslint": "^7.0.0",
+    "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-jsx-a11y": "^6.2.3",
     "eslint-plugin-prefer-object-spread": "^1.2.1",
-    "eslint-plugin-prettier": "^3.1.2",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^2.5.1",
-    "null-loader": "^3.0.0",
-    "prettier": "^2.0.2",
-    "typescript": "^3.8.3"
+    "eslint-plugin-prettier": "^3.1.3",
+    "eslint-plugin-react": "^7.20.0",
+    "eslint-plugin-react-hooks": "^4.0.2",
+    "null-loader": "^4.0.0",
+    "prettier": "^2.0.5",
+    "typescript": "^3.9.3"
   }
 }
