--- src/components/chrome/_index.scss
@@ -1,1 +1,0 @@
-@import 'chrome';
