+++ src/components/navigation_links/admin_links.ts
@@ -1,4 +1,8 @@
-export const buildAdminLinks = (makeAction: (path: string) => () => void) => [
+import { EuiNavDrawerGroupProps } from '@elastic/eui';
+
+export const buildAdminLinks = (
+  makeAction: (path: string) => () => void
+): EuiNavDrawerGroupProps['listItems'] => [
   {
     label: 'Admin',
     iconType: 'managementApp',
