+++ src/components/navigation_links/pin_extra_action.ts
@@ -1,4 +1,6 @@
-export const pinExtraAction = {
+import { EuiButtonIconPropsForButton } from '@elastic/eui';
+
+export const pinExtraAction: EuiButtonIconPropsForButton = {
   color: 'subdued',
   iconType: 'pin',
   iconSize: 's',
